List Part 2 -> Number Made More Robust
The Number Class can now take in any type. This in Preperation take in List of List, and Custom Type here forward.  Now List can take in a vector of Numbers.
List can now Add, Subtract, query and Concatenate other list.
List will be the bases of my grid so this was essential.